+++ data/statistics_x01.go
@@ -187,12 +187,7 @@ func GetPlayerX01PreviousStatistics(id int) (*models.StatisticsX01, error) {
 		return nil, err
 	}
 	if len(statistics) > 0 {
-		stats := statistics[0]
-		if err != nil {
-			return nil, err
-		}
-
-		return stats, nil
+		return statistics[0], nil
 	}
 	return new(models.StatisticsX01), nil
 }
@@ -230,8 +225,8 @@ func GetPlayersX01Statistics(ids []int, startingScores ...int) ([]*models.Statis
 			LEFT JOIN matches m2 ON m2.id = l2.match_id AND l2.winner_id = p.id
 		WHERE s.player_id IN (?)
 			AND l.starting_score IN (?)
-			AND l.is_finished = 1 AND m.is_abandoned = 0 AND m.is_practice = 0 AND m.is_walkover = 0
-			AND IFNULL(l.leg_type_id, m.match_type_id) = 1
+			AND l.is_finished = 1 AND m.is_abandoned = 0 AND m.is_walkover = 0
+			AND COALESCE(l.leg_type_id, m.match_type_id) = 1
 		GROUP BY s.player_id
 		ORDER BY p.id`, ids, startingScores)
 	if err != nil {
@@ -318,8 +313,8 @@ func GetPlayersX01PreviousStatistics(ids []int, startingScores ...int) ([]*model
 			LEFT JOIN matches m2 ON m2.id = l2.match_id AND l2.winner_id = p.id
 		WHERE s.player_id IN (?)
 			AND l.starting_score IN (?)
-			AND l.is_finished = 1 AND m.is_abandoned = 0 AND m.is_practice = 0 AND m.is_walkover = 0
-			AND m.match_type_id = 1
+			AND l.is_finished = 1 AND m.is_abandoned = 0 AND m.is_walkover = 0
+			AND COALESCE(l.leg_type_id, m.match_type_id) = 1
 			-- Exclude all matches played this week
 			AND m.updated_at < (CURRENT_DATE - INTERVAL WEEKDAY(CURRENT_DATE) DAY)
 		GROUP BY s.player_id
@@ -464,15 +459,6 @@ func GetX01StatisticsForPlayer(id int, matchType int) (*models.StatisticsX01, er
 
 // GetX01HistoryForPlayer will return history of X01 statistics for the given player
 func GetX01HistoryForPlayer(id int, limit int, matchType int) ([]*models.Leg, error) {
-	legs, err := GetLegsOfType(matchType, false)
-	if err != nil {
-		return nil, err
-	}
-	m := make(map[int]*models.Leg)
-	for _, leg := range legs {
-		m[leg.ID] = leg
-	}
-
 	rows, err := models.DB.Query(`
 		SELECT
 			l.id AS 'leg_id',
@@ -506,7 +492,8 @@ func GetX01HistoryForPlayer(id int, limit int, matchType int) ([]*models.Leg, er
 	}
 	defer rows.Close()
 
-	legs = make([]*models.Leg, 0)
+	statistics := make(map[int]*models.StatisticsX01, 0)
+	legIDs := make([]int, 0)
 	for rows.Next() {
 		s := new(models.StatisticsX01)
 		err := rows.Scan(&s.LegID, &s.PlayerID, &s.PPD, &s.FirstNinePPD, &s.ThreeDartAvg, &s.FirstNineThreeDartAvg, &s.Score60sPlus, &s.Score100sPlus,
@@ -515,9 +502,19 @@ func GetX01HistoryForPlayer(id int, limit int, matchType int) ([]*models.Leg, er
 		if err != nil {
 			return nil, err
 		}
-		leg := m[s.LegID]
-		leg.Statistics = s
-		legs = append(legs, leg)
+		statistics[s.LegID] = s
+		legIDs = append(legIDs, s.LegID)
+	}
+
+	if len(legIDs) == 0 {
+		return []*models.Leg{}, nil
+	}
+	legs, err := GetLegs(legIDs)
+	if err != nil {
+		return nil, err
+	}
+	for _, leg := range legs {
+		leg.Statistics = statistics[leg.ID]
 	}
 	return legs, nil
 }
@@ -924,7 +921,7 @@ func RecalculateX01Statistics(legs []int) ([]string, error) {
 }
 
 // GetPlayerBadgeStatistics will return statistics calculate badges for the given players
-func GetPlayerBadgeStatistics(ids []int, legID *int) (map[int]*models.BadgeStatistics, error) {
+func GetPlayerBadgeStatistics(ids []int, legID *int) (map[int]*models.PlayerBadgeStatistics, error) {
 	q, args, err := sqlx.In(`
 		SELECT
 			player_id,
@@ -934,7 +931,8 @@ func GetPlayerBadgeStatistics(ids []int, legID *int) (map[int]*models.BadgeStati
 		FROM statistics_x01 s
 			LEFT JOIN leg l ON s.leg_id = l.id
 			LEFT JOIN matches m ON l.match_id = m.id
-		WHERE player_id IN (?) AND l.num_players = 2 AND m.is_practice = 0
+		WHERE player_id IN (?)
+			AND l.is_finished = 1 AND m.is_abandoned = 0 AND m.is_walkover = 0
 			AND m.is_bye = 0 AND m.is_walkover = 0 AND leg_id <= COALESCE(?, ~0) -- BIGINT hack
 		GROUP BY player_id
 		ORDER BY player_id DESC`, ids, legID)
@@ -947,12 +945,13 @@ func GetPlayerBadgeStatistics(ids []int, legID *int) (map[int]*models.BadgeStati
 	}
 	defer rows.Close()
 
-	statistics := make(map[int]*models.BadgeStatistics)
+	statistics := make(map[int]*models.PlayerBadgeStatistics)
 	for _, playerID := range ids {
-		statistics[playerID] = new(models.BadgeStatistics)
+		statistics[playerID] = new(models.PlayerBadgeStatistics)
 	}
 	for rows.Next() {
-		s := new(models.BadgeStatistics)
+		s := new(models.PlayerBadgeStatistics)
+		s.BadgeMap = make(map[int]interface{}, 0)
 		err := rows.Scan(&s.PlayerID, &s.Score100sPlus, &s.Score140sPlus, &s.Score180s)
 		if err != nil {
 			return nil, err
@@ -963,5 +962,52 @@ func GetPlayerBadgeStatistics(ids []int, legID *int) (map[int]*models.BadgeStati
 		return nil, err
 	}
 
+	q, args, err = sqlx.In(`
+		SELECT
+			player_id, first_dart
+		FROM score s
+			LEFT JOIN leg l ON s.leg_id = l.id
+			LEFT JOIN matches m ON l.match_id = m.id
+		WHERE
+			first_dart = second_dart AND first_dart = third_dart
+			and ((first_dart_multiplier = 1 and second_dart_multiplier = 2 and third_dart_multiplier = 3)
+				or (first_dart_multiplier = 2 and second_dart_multiplier = 3 and third_dart_multiplier = 1)
+				or (first_dart_multiplier = 3 and second_dart_multiplier = 1 and third_dart_multiplier = 2)
+				or (first_dart_multiplier = 3 and second_dart_multiplier = 2 and third_dart_multiplier = 1)
+				or (first_dart_multiplier = 1 and second_dart_multiplier = 3 and third_dart_multiplier = 2)
+				or (first_dart_multiplier = 2 and second_dart_multiplier = 1 and third_dart_multiplier = 3))
+			AND is_bust = 0 AND s.player_id IN (?) AND leg_id <= COALESCE(?, ~0) -- BIGINT hack
+			AND l.is_finished = 1 AND m.is_abandoned = 0 AND m.is_walkover = 0
+			AND m.is_bye = 0 AND m.is_walkover = 0
+		GROUP BY first_dart, player_id`, ids, legID)
+	if err != nil {
+		return nil, err
+	}
+	rows, err = models.DB.Query(q, args...)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+
+	shanghai := make(map[int][]int, 0)
+	for _, playerID := range ids {
+		shanghai[playerID] = make([]int, 0)
+	}
+	for rows.Next() {
+		var playerID int
+		var number int
+		err := rows.Scan(&playerID, &number)
+		if err != nil {
+			return nil, err
+		}
+		shanghai[playerID] = append(shanghai[playerID], number)
+	}
+	if err = rows.Err(); err != nil {
+		return nil, err
+	}
+
+	for playerID, stats := range statistics {
+		stats.Shanghais = shanghai[playerID]
+	}
 	return statistics, nil
 }