+++ models/visit.go
@@ -9,6 +9,9 @@ import (
 	"github.com/guregu/null"
 )
 
+var NUMS_WHITE = []int{0, 1, 4, 6, 15, 17, 19, 16, 11, 9, 5}
+var NUMS_BLACK = []int{0, 20, 18, 13, 10, 2, 3, 7, 8, 14, 12}
+
 // Visit struct used for storing legs
 type Visit struct {
 	ID          int         `json:"id"`
@@ -441,13 +444,13 @@ func (visit *Visit) CalculateAroundTheClockScore(currentScore int) int {
 func (visit *Visit) CalculateAroundTheWorldScore(round int) int {
 	score := 0
 	if round == visit.FirstDart.ValueRaw() || (round == 21 && visit.FirstDart.IsBull()) {
-		score += visit.FirstDart.GetScore()
+		score += int(visit.FirstDart.Multiplier)
 	}
 	if round == visit.SecondDart.ValueRaw() || (round == 21 && visit.SecondDart.IsBull()) {
-		score += visit.SecondDart.GetScore()
+		score += int(visit.SecondDart.Multiplier)
 	}
 	if round == visit.ThirdDart.ValueRaw() || (round == 21 && visit.ThirdDart.IsBull()) {
-		score += visit.ThirdDart.GetScore()
+		score += int(visit.ThirdDart.Multiplier)
 	}
 	return score
 }
@@ -599,6 +602,28 @@ func (visit *Visit) CalculateScamMarks(scores map[int]*Player2Leg) int {
 	hits.Add(visit.ThirdDart)
 
 	return marks
+}
+
+func (visit *Visit) Calculate170Score(round int, player *Player2Leg) int {
+	player.DartsThrown += 3
+	score := 0
+	if !visit.IsBust {
+		player.CurrentScore -= visit.GetScore()
+		score = visit.GetScore()
+	}
+
+	if player.CurrentScore == 0 && visit.GetLastDart().IsDouble() {
+		// We hit a checkout, reset
+		player.CurrentScore = 170
+		player.CurrentPoints.Int64++
+		player.DartsThrown = 0
+	} else if round != 1 && player.DartsThrown%9 == 0 {
+		// 9 Darts have been thrown, reset
+		player.CurrentScore = 170
+		player.DartsThrown = 0
+		score = 0
+	}
+	return score
 }
 
 // IsShanghai will check if the given visit is a "Shanghai". A Shanghai visit is one where a single, double and triple multipler is hit with each dart